fixup! Improve user deletion - clean all personal data

+++ app/models/person/concerns/delete_person.rb
@@ -32,7 +32,7 @@ module DeletePerson
           person.emails.destroy_all
 
           # Delete location
-          person.location.destroy
+          person.location.destroy if person.location
 
           # Delete avatar
           person.image.destroy
